TEST-33 Relocated EmptyStatement to make it reusable
