+++ dominion-pax-exam-invokers/src/main/java/org/codice/dominion/pax/exam/invoker/DominionParameterizedContainerTestRunner.java
@@ -14,6 +14,7 @@
 package org.codice.dominion.pax.exam.invoker;
 
 import java.util.List;
+import org.codice.junit.rules.EmptyStatement;
 import org.codice.junit.rules.MethodRuleAnnotationProcessor;
 import org.junit.rules.MethodRule;
 import org.junit.runner.notification.RunNotifier;
