+++ junit-extensions/src/main/java/org/codice/junit/rules/EmptyStatement.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.dominion.pax.exam.invoker;
+package org.codice.junit.rules;
 
 import org.junit.runners.model.Statement;
